:recycle: Integrate safer defaults + add AppSettings into CurrencyListVM

+++ Walut/Managers/SwiftDataManager.swift
@@ -10,10 +10,10 @@ import SwiftData
 
 struct SwiftDataManager {
     @MainActor
-    static func saveCurrencies(data: [Currency], base: String = SharedDataManager.shared.base.code, to modelContext: ModelContext) {
+    static func saveCurrencies(data: [Currency], base: String = Defaults.baseCode()!, to modelContext: ModelContext) {
         do {
             let savedCurrencies = try modelContext.fetch(FetchDescriptor<SavedCurrency>())
-            let nextUpdate = UserDefaults(suiteName: "group.dev.bartminski.Walut")!.integer(forKey: "nextUpdate")
+            let nextUpdate = Defaults.nextUpdateInterval()
             
             data.forEach { item in
                 if !savedCurrencies.contains(where: { $0.code == item.code && $0.base == base && $0.nextRefresh == nextUpdate }) {
@@ -27,17 +27,17 @@ struct SwiftDataManager {
         }
     }
     
-    static func getCurrencies(from modelContext: ModelContext) -> [Currency] {
+    @MainActor
+    static func getCurrencies(from modelContext: ModelContext, baseCode: String) -> [Currency] {
         do {
             let savedCurrencies = try modelContext.fetch(FetchDescriptor<SavedCurrency>())
-            let shared = SharedDataManager.shared
-            let nextUpdate = UserDefaults(suiteName: "group.dev.bartminski.Walut")!.integer(forKey: "nextUpdate")
+            let nextUpdate = Defaults.nextUpdateInterval()
             
             let currencies = savedCurrencies
-                .filter { $0.nextRefresh == nextUpdate && $0.base == shared.base.code }
+                .filter { $0.nextRefresh == nextUpdate && $0.base == baseCode }
                 .map { Currency(from: $0) }
                 .map { currency in
-                    let lastRate = getLastRate(for: currency.code, base: shared.base.code, from: savedCurrencies)
+                    let lastRate = getLastRate(for: currency.code, base: baseCode, from: savedCurrencies)
                     
                     if let lastRate {
                         var newCurrency = currency
@@ -55,12 +55,13 @@ struct SwiftDataManager {
     }
     
     @MainActor
-    static func cleanData(from modelContext: ModelContext, useInterval: Bool = false) {
+    static func cleanData(from modelContext: ModelContext, useInterval: Bool = false, storageOption: StorageSavingOptions? = nil) {
         do {
-            let nextUpdate = UserDefaults(suiteName: "group.dev.bartminski.Walut")!.integer(forKey: "nextUpdate")
+            let nextUpdate = Defaults.nextUpdateInterval()
             
             if useInterval {
-                let minInterval = decodeStorageOptionInterval()
+                guard let storageOption else { print("No storage option provided!"); return }
+                let minInterval = decodeStorageOptionInterval(from: storageOption)
                 try modelContext.delete(model: SavedCurrency.self, where: #Predicate {
                     $0.nextRefresh < minInterval || $0.nextRefresh > nextUpdate - (12 * 3600) && $0.nextRefresh < nextUpdate
                 })
@@ -76,9 +77,9 @@ struct SwiftDataManager {
         }
     }
     
-    static func decodeStorageOptionInterval() -> Int {
+    static func decodeStorageOptionInterval(from option: StorageSavingOptions) -> Int {
         var timestamp = Date().timeIntervalSince1970
-        switch SharedDataManager.shared.storageOption {
+        switch option {
         case .twoDays:
             timestamp -= 86400 * 2
         case .oneWeek:
@@ -99,7 +100,7 @@ struct SwiftDataManager {
     static func getWidgetData(for code: String, base: String, from modelContext: ModelContext) -> Currency? {
         let descriptor = FetchDescriptor<SavedCurrency>()
         let saved = try? modelContext.fetch(descriptor)
-        let nextUpdate = UserDefaults(suiteName: "group.dev.bartminski.Walut")!.integer(forKey: "nextUpdate")
+        let nextUpdate = Defaults.nextUpdateInterval()
         let currency = saved?
             .filter { $0.nextRefresh == nextUpdate && $0.base == base }
             .map { Currency(from: $0) }
